dragon: DragonKeyboardFirmwareUpdater: Change the life cycle of the service
The original life cycle of the service is running on boot completed and running
persistently in the background. This CL changes the event for starting the
service from boot completed to Bluetooth device connected. Except for the
period of checking the firmware update, waiting for user's confirmation and
pushing the new firmware, the service should not be running in the rest of the
time.
This CL also defines a postpone time (a week) if the user clicks on LATER button
on the update notification.

Bug: 25369339
TEST=build, run the service to update the keyboard firmware and monitor its
status

Change-Id: I9e9170c0dbda5ba937fce585fb4899418101ca15