audio sound_trigger: Check null pointer in streaming
A case was encountered at resume that recognition callback got
triggered but failed to read more data from the DSP. Current
error handling closes the pcm for this error but never open it
until start_recognition is called again. This causes media server
crash at the next attempt to read samples from the NULL pcm.
Add NULL checks for this pcm so that this failure can be handled
by the caller instead of crash.

BUG=25875632

Change-Id: Ic4817baada45560856fc26c5e8fe89facb773696