Enable 4-channel input
Smaug has 4 microphones. Enable 4-channel input that will be processed to
2-channel by a noise-suppression module.

Kernel ALSA does not like when when open codec (in case of Smaug rt5677)
both for input and output with different format (2-channel TDM vs 4-channel TDM).
It confuses ALSA and it gets dropping half of all frames.

To avoid that we open output with 4-channel format as well and scatter data
2->4 channel somewhat reverse what we do at the input.

Change-Id: I41205f1231b721aed0ad35ecb7eb895955cd5ef9